+++ service/controller/app/v1/key/key.go
@@ -2,6 +2,8 @@ package key
 
 import (
 	"fmt"
+	"net/url"
+	"path"
 
 	"github.com/giantswarm/apiextensions/pkg/apis/application/v1alpha1"
 	"github.com/giantswarm/microerror"
@@ -156,4 +158,16 @@ func VersionLabel(customResource v1alpha1.App) string {
 	} else {
 		return ""
 	}
+}
+
+func GenerateTarballURL(baseURL string, appName string, version string) (string, error) {
+	if baseURL == "" || appName == "" || version == "" {
+		return "", microerror.Maskf(executionFailedError, "baseURL %#q, appName %#q, release %#q should not be empty", baseURL, appName, version)
+	}
+	u, err := url.Parse(baseURL)
+	if err != nil {
+		return "", microerror.Mask(err)
+	}
+	u.Path = path.Join(u.Path, fmt.Sprintf("%s-%s.tgz", appName, version))
+	return u.String(), nil
 }
